xuat hang - dx kh bdg

+++ src/main/java/com/tcdt/qlnvhang/service/xuathang/xuattheophuongthucdaugia/tochuctrienkhai/ketqua/XhKqBdgHdrServiceImpl.java
@@ -1,9 +1,13 @@
 package com.tcdt.qlnvhang.service.xuathang.xuattheophuongthucdaugia.tochuctrienkhai.ketqua;
 
+import com.tcdt.qlnvhang.enums.NhapXuatHangTrangThaiEnum;
 import com.tcdt.qlnvhang.repository.xuathang.xuattheophuongthucdaugia.tochuctrienkhai.ketqua.XhKqBdgHdrRepository;
 import com.tcdt.qlnvhang.request.xuathang.xuattheophuongthucdaugia.tochuctrienkhai.ketqua.XhKqBdgHdrReq;
+import com.tcdt.qlnvhang.service.SecurityContextService;
 import com.tcdt.qlnvhang.service.impl.BaseServiceImpl;
+import com.tcdt.qlnvhang.table.UserInfo;
 import com.tcdt.qlnvhang.table.xuathang.xuattheophuongthucdaugia.tochuctrienkhai.ketqua.XhKqBdgHdr;
+import com.tcdt.qlnvhang.util.Contains;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -12,6 +16,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.util.ObjectUtils;
+import org.springframework.util.StringUtils;
 
 import javax.servlet.http.HttpServletResponse;
 import java.time.LocalDate;
@@ -81,7 +86,45 @@ public class XhKqBdgHdrServiceImpl extends BaseServiceImpl implements XhKqBdgHdr
 
     @Override
     public XhKqBdgHdr approve(XhKqBdgHdrReq req) throws Exception {
-        return null;
+        UserInfo userInfo= SecurityContextService.getUser();
+        if (userInfo == null){
+            throw new Exception("Bad request.");
+        }
+
+        if (StringUtils.isEmpty(req.getId())){
+            throw new Exception("Không tìm thấy dữ liệu");
+        }
+
+        Optional<XhKqBdgHdr> optional = xhKqBdgHdrRepository.findById(req.getId());
+        if(!optional.isPresent()){
+            throw new Exception("Không tìm thấy dữ liệu");
+        }
+
+        XhKqBdgHdr data = optional.get();
+        String status = req.getTrangThai() + data.getTrangThai();
+        switch (status) {
+            case Contains.CHODUYET_TP + Contains.DUTHAO:
+            case Contains.CHODUYET_TP + Contains.TUCHOI_TP:
+            case Contains.CHODUYET_TP + Contains.TUCHOI_LDC:
+                data.setNguoiGuiDuyetId(userInfo.getId());
+                data.setNgayGuiDuyet(getDateTimeNow());
+            case Contains.TUCHOI_TP + Contains.CHODUYET_TP:
+            case Contains.TUCHOI_LDC + Contains.CHODUYET_LDC:
+                data.setNguoiPduyetId(userInfo.getId());
+                data.setNgayPduyet(getDateTimeNow());
+                data.setLyDoTuChoi(req.getLyDoTuChoi());
+                break;
+            case Contains.CHODUYET_LDC + Contains.CHODUYET_TP:
+            case Contains.DADUYET_LDC + Contains.CHODUYET_LDC:
+            case Contains.BAN_HANH + Contains.DADUYET_LDC:
+                data.setNguoiPduyetId(userInfo.getId());
+                data.setNgayPduyet(getDateTimeNow());
+                break;
+            default:
+                throw new Exception("Phê duyệt không thành công");
+        }
+        data.setTrangThai(req.getTrangThai());
+        return xhKqBdgHdrRepository.save(data);
     }
 
     @Override
@@ -92,6 +135,9 @@ public class XhKqBdgHdrServiceImpl extends BaseServiceImpl implements XhKqBdgHdr
         Optional<XhKqBdgHdr> byId = xhKqBdgHdrRepository.findById(id);
         if (!byId.isPresent()) {
             throw new Exception("Không tìm thấy dữ liệu");
+        }
+        if(!byId.get().getTrangThai().equals(NhapXuatHangTrangThaiEnum.DUTHAO.getId())){
+            throw new Exception("Chỉ được xóa bản ghi với trang thái là dự thảo");
         }
         xhKqBdgHdrRepository.delete(byId.get());
     }
